moved index types into their own file, fixes circular deps, cleaner placement
