+++ src/types/index-types.ts
@@ -1,0 +1,41 @@
+export type QueryOptions = {
+  sortProperty: string;
+  sortDirection?: SortDirection;
+  limit?: number;
+  cursor?: string;
+};
+
+export enum SortDirection {
+  Descending = -1,
+  Ascending = 1
+}
+
+export type EqualFilter = string | number | boolean;
+
+export type OneOfFilter = EqualFilter[];
+
+export type RangeValue = string | number;
+
+/**
+ * "greater than" or "greater than or equal to" range condition. `gt` and `gte` are mutually exclusive.
+ */
+export type GT = ({ gt: RangeValue } & { gte?: never }) | ({ gt?: never } & { gte: RangeValue });
+
+/**
+ * "less than" or "less than or equal to" range condition. `lt`, `lte` are mutually exclusive.
+ */
+export type LT = ({ lt: RangeValue } & { lte?: never }) | ({ lt?: never } & { lte: RangeValue });
+
+/**
+ * Ranger filter. 1 condition is required.
+ */
+export type RangeFilter = (GT | LT) & Partial<GT> & Partial<LT>;
+
+export type FilterValue = EqualFilter | OneOfFilter | RangeFilter;
+
+export type Filter = {
+  [property: string]: FilterValue;
+};
+
+export type Indexes = { [key: string]: unknown };
+export type IndexedItem<T> = { itemId: string, value: T, indexes: Indexes };
