+++ src/types/message-types.ts
@@ -1,5 +1,6 @@
 import type { DelegatedGrantMessage } from '../types/delegated-grant-message.js';
 import type { GeneralJws } from './jws-types.js';
+import type { SortDirection } from './index-types.js';
 
 /**
  * Intersection type for all concrete message types.
@@ -68,36 +69,6 @@ export type QueryResultEntry = {
   encodedData?: string;
 };
 
-export type EqualFilter = string | number | boolean;
-
-export type OneOfFilter = EqualFilter[];
-
-export type RangeValue = string | number;
-
-/**
- * "greater than" or "greater than or equal to" range condition. `gt` and `gte` are mutually exclusive.
- */
-export type GT = ({ gt: RangeValue } & { gte?: never }) | ({ gt?: never } & { gte: RangeValue });
-
-/**
- * "less than" or "less than or equal to" range condition. `lt`, `lte` are mutually exclusive.
- */
-export type LT = ({ lt: RangeValue } & { lte?: never }) | ({ lt?: never } & { lte: RangeValue });
-
-/**
- * Ranger filter. 1 condition is required.
- */
-export type RangeFilter = (GT | LT) & Partial<GT> & Partial<LT>;
-
-export type FilterValue = EqualFilter | OneOfFilter | RangeFilter;
-
-export type Filter = {
-  [property: string]: FilterValue;
-};
-
-export type Indexes = { [key: string]: unknown };
-export type IndexedItem<T> = { itemId: string, value: T, indexes: Indexes };
-
 /**
  * Pagination Options for querying messages.
  *
@@ -108,10 +79,7 @@ export type Pagination = {
   limit?: number;
 };
 
-export enum SortDirection {
-  Descending = -1,
-  Ascending = 1
-}
+
 
 export type MessageSort = {
   dateCreated?: SortDirection;
